nerves_key: add long serial number support
This adds long serial number support based on the OTP data's flags
setting. The code has been written so that if the flag has not been set,
`boardid` only makes one read request to the ATECC. If the flag is set,
then two reads are made so that the remainder of the serial number can
be read. This was done in an abundance of caution to avoid changing any
I2C bus traffic on production devices.