rpi_eth0/wlan0: call vc directly and autodetect
This removes the exec of vcmailbox and the need for rpi-userland to be
installed. The ioctl interface turned out to be nearly trivial. The
downside is that it's no longer possible to unit test with the current
available test infrastructure.

This also removes the RPi4 variant of the options. This can be detected
by OUI and in all likelihood the next Raspberry Pi's may use the newer
approach (add 1 instead of XOR by 0x555555) anyway.