Improve the ATECC retry logic
This fixes an issue where a failed ATECC I2C read would not retry. It
also doubles the number of retries since they don't take a long time
and the consequences of not reading a ATECC serial number through a
momentary hiccup are high. Lastly, it raises the retry timeout from
1.5 ms to 10 ms to give the system a little more time to settle if
something is happening.

When out of retries, it prints an error message that will hopefully aid
debugging. Errors on retries aren't printed any more, since they look
scary and sometimes are due to two processes trying to talk to the ATECC
at the same time.