Prefer the sysfs path for reading device tree serials
The /proc/device-tree/serial-number path is a symlink to the sysfs
path, so this really just saves following a symlink. If the sysfs path
somehow doesn't exist, the old /proc location is tried.

This change won't break anything and I bet the /proc location won't be
removed. It looks better to me to try the preferred location first.